backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
openai==0.27.8
langchain==0.0.235
redis==4.6.0
pydantic==1.10.7
python-multipart==0.0.6
sentry-sdk[fastapi]==1.25.1
prometheus-fastapi-instrumentator==6.3.0
pytest==7.3.1
httpx==0.24.1
